variable "n" (day of the year) changed from float to integer (again) in documentation

+++ solar_radiation.py
@@ -16,7 +16,7 @@ def check_nth_day_range(n):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
 
     Returns
@@ -101,7 +101,7 @@ def B_nth_day(n):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
 
     Returns
@@ -121,7 +121,7 @@ def Gon(n):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
 
     Returns
@@ -142,7 +142,7 @@ def Eq_time(n):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
 
     Returns
@@ -164,7 +164,7 @@ def declination(n):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
 
     Returns
@@ -186,7 +186,7 @@ def solar_time(n, hour, minute, lng):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     hour : integer
         hour of the day (0 to 23)
@@ -260,7 +260,7 @@ def theta(n, lat, beta, surf_az, hour, minute):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -307,7 +307,7 @@ def theta_z(n, lat, hour, minute):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -342,7 +342,7 @@ def solar_azimuth(n, lat, hour, minute):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -391,7 +391,7 @@ def solar_altitude(n, lat, hour, minute):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -427,7 +427,7 @@ def sunset_hour_angle(n, lat):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -455,7 +455,7 @@ def sunset_time(n, lat):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -486,7 +486,7 @@ def sunrise_hour_angle(n, lat):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -506,7 +506,7 @@ def sunrise_time(n, lat):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -537,7 +537,7 @@ def daylight_hours(n, lat):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
@@ -651,7 +651,7 @@ def solar_vector_NED(n, lat, hour, minute):
 
     Parameters
     ----------
-    n : float
+    n : integer
         day of the year (1 to 365)
     lat : float
         latitude (-90 to 90) in degrees
